refactor(cli): fold TestPlanDescriptor into SideEffectDescriptor
Drops the parallel \`testPlanDescriptors\` field on Project/Workspace
descriptors in favour of a new \`SideEffectDescriptor.testPlan\` case.
TestPlanDescriptor moves to TuistCore next to the other descriptors and
grows an \`encode()\` method; SideEffectDescriptorExecutor handles the new
case by writing the encoded JSON at execution time. Because side effects
run after \`xcodeProj.write()\` in XcodeProjWriter, the PBX blueprint
identifiers embedded in the generated \`.xctestplan\` are already stable
by the time the descriptor is encoded, so no separate ordering hook is
needed.
