+++ cli/Sources/TuistGenerator/Descriptors/ProjectDescriptor.swift
@@ -27,24 +27,18 @@ public struct ProjectDescriptor {
     /// The side effects required for generating this project
     public var sideEffectDescriptors: [SideEffectDescriptor]
 
-    /// Test plans whose `.xctestplan` files are generated by Tuist.
-    /// Written after the `.xcodeproj` so that target blueprint identifiers are stable.
-    public var testPlanDescriptors: [TestPlanDescriptor]
-
     public init(
         path: AbsolutePath,
         xcodeprojPath: AbsolutePath,
         xcodeProj: XcodeProj,
         schemeDescriptors: [SchemeDescriptor],
-        sideEffectDescriptors: [SideEffectDescriptor],
-        testPlanDescriptors: [TestPlanDescriptor] = []
+        sideEffectDescriptors: [SideEffectDescriptor]
     ) {
         self.path = path
         self.xcodeprojPath = xcodeprojPath
         self.xcodeProj = xcodeProj
         self.schemeDescriptors = schemeDescriptors
         self.sideEffectDescriptors = sideEffectDescriptors
-        self.testPlanDescriptors = testPlanDescriptors
     }
 
     #if DEBUG
