+++ cli/Sources/TuistGenerator/Descriptors/WorkspaceDescriptor.swift
@@ -32,9 +32,6 @@ public struct WorkspaceDescriptor {
     /// The side effects required for generating this workspace
     public var sideEffectDescriptors: [SideEffectDescriptor]
 
-    /// Test plans declared at the workspace level whose `.xctestplan` files are generated by Tuist.
-    public var testPlanDescriptors: [TestPlanDescriptor]
-
     /// The descriptor used to generate workspace settings (WorkspaceSettings.xcsettings)
     public var workspaceSettingsDescriptor: WorkspaceSettingsDescriptor?
 
@@ -45,7 +42,6 @@ public struct WorkspaceDescriptor {
         projectDescriptors: [ProjectDescriptor],
         schemeDescriptors: [SchemeDescriptor],
         sideEffectDescriptors: [SideEffectDescriptor],
-        testPlanDescriptors: [TestPlanDescriptor] = [],
         workspaceSettingsDescriptor: WorkspaceSettingsDescriptor? = nil
     ) {
         self.path = path
@@ -55,7 +51,6 @@ public struct WorkspaceDescriptor {
         self.projectDescriptors = projectDescriptors
         self.schemeDescriptors = schemeDescriptors
         self.sideEffectDescriptors = sideEffectDescriptors
-        self.testPlanDescriptors = testPlanDescriptors
     }
 
     #if DEBUG
