+++ cli/Sources/TuistGenerator/Generator/TestPlanGenerator.swift
@@ -2,12 +2,12 @@ import Foundation
 import Path
 import TuistCore
 import XcodeGraph
-import XcodeProj
 
-/// Builds `TestPlanDescriptor` values from the graph's generated test plans and the generated projects that own their targets.
+/// Builds `TestPlanDescriptor` values from the graph's generated test plans and the generated
+/// projects that own their targets.
 enum TestPlanGenerator {
-    /// Produces a descriptor for a generated test plan, returning `nil` when none of its test targets can be resolved
-    /// against the generated projects.
+    /// Produces a descriptor for a generated test plan, returning `nil` when none of its test
+    /// targets can be resolved against the generated projects.
     static func descriptor(
         for testPlan: TestPlan,
         graphTraverser: GraphTraversing,
@@ -44,60 +44,4 @@ enum TestPlanGenerator {
 
         return TestPlanDescriptor(path: testPlan.path, testTargets: testTargets)
     }
-
-    /// Encodes a `TestPlanDescriptor` into the Xcode `.xctestplan` JSON format.
-    ///
-    /// - Note: Must be called after the owning `.xcodeproj` has been written so that `pbxTarget.uuid`
-    ///   returns stable blueprint identifiers.
-    static func encode(_ descriptor: TestPlanDescriptor) throws -> Data {
-        let plan = XCTestPlanPayload(
-            configurations: [
-                XCTestPlanPayload.Configuration(
-                    id: UUID(uuidString: "91BDB644-1AEA-4734-9E55-F6DA2F59DF74")!,
-                    name: "Configuration 1",
-                    options: [:]
-                ),
-            ],
-            defaultOptions: [:],
-            testTargets: descriptor.testTargets.map { target in
-                XCTestPlanPayload.TestTarget(
-                    enabled: target.isEnabled ? nil : false,
-                    target: XCTestPlanPayload.TargetReference(
-                        containerPath: target.containerPath,
-                        identifier: target.pbxTarget.uuid,
-                        name: target.pbxTarget.name
-                    )
-                )
-            },
-            version: 1
-        )
-
-        let encoder = JSONEncoder()
-        encoder.outputFormatting = [.prettyPrinted, .sortedKeys, .withoutEscapingSlashes]
-        return try encoder.encode(plan)
-    }
-}
-
-private struct XCTestPlanPayload: Encodable {
-    struct Configuration: Encodable {
-        let id: UUID
-        let name: String
-        let options: [String: String]
-    }
-
-    struct TargetReference: Encodable {
-        let containerPath: String
-        let identifier: String
-        let name: String
-    }
-
-    struct TestTarget: Encodable {
-        let enabled: Bool?
-        let target: TargetReference
-    }
-
-    let configurations: [Configuration]
-    let defaultOptions: [String: String]
-    let testTargets: [TestTarget]
-    let version: Int
 }
